requirements.txt
pyyaml
diskimage-builder
python-openstackclient
python-glanceclient
